# openapi.yaml
openapi: 3.0.3
info:
  title: QA Retail — API Backend
  description: >
    Contrato de la API REST de "ChileRetail", la tienda demo usada como campo de
    practica de QA. Documenta, en formato contract-first (OpenAPI), el
    comportamiento real de cada endpoint ya implementado en `src/routes/*.routes.js`:
    parametros, cuerpos de peticion, respuestas y codigos de estado (incluyendo los
    casos negativos sembrados a proposito para practicas de testing).

    Reglas transversales:
    - Todas las rutas estan bajo el prefijo `/api` (ver `servers`).
    - Las rutas marcadas con el candado requieren `Authorization: Bearer <token>`
      (JWT, expira en 2h). Sin token o token invalido/expirado -> **401**.
    - Las rutas marcadas ademas como "admin" requieren `role: admin` en el token.
      Autenticado pero sin rol admin -> **403**.
    - Toda ruta no definida en este contrato responde **404**
      `{ "error": "Ruta no encontrada" }`.
    - Cualquier excepcion no controlada responde **500**
      `{ "error": "Error interno del servidor" }`.
  version: "1.0.0"
  contact:
    name: QA Retail — Backend
servers:
  - url: http://localhost:4000/api
    description: Servidor local de desarrollo
...
security: []
paths:
...
  /cart:
    get:
      tags: [cart]
      summary: Carro del usuario autenticado
      description: >
        Cada item viene "enriquecido" con el producto completo y `lineTotal`
        (`price * quantity`). Si el producto fue eliminado del catalogo,
        `product` es `null` y `lineTotal` es `0`.
      operationId: getCart
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Contenido del carro.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
  /cart/items:
    post:
      tags: [cart]
      summary: Agrega un producto al carro (o acumula cantidad si ya existe)
      operationId: addCartItem
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCartItemRequest"
      responses:
        "201":
          description: Item agregado; se retorna el carro completo.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartResponse"
        "400":
          description: Faltan `productId`/`quantity`, o `quantity` no es mayor que 0.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: >
            Stock insuficiente: la cantidad acumulada (existente + solicitada)
            supera el `stock` disponible del producto.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/ServerError"
  /cart/items/{productId}:
    parameters:
      - name: productId
        in: path
        required: true
        schema: { type: integer }
        example: 1
    put:
      tags: [cart]
      summary: Fija la cantidad de un item existente en el carro
      description: >
        No valida contra el `stock` disponible (a diferencia de `POST /cart/items`).
        Para eliminar el item usar `DELETE` en vez de `quantity: 0`.
      operationId: setCartItemQuantity
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetCartItemQuantityRequest"
      responses:
        "200":
          description: Carro actualizado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartResponse"
        "400":
          description: "`quantity` ausente o no mayor que 0."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: El producto no existe, o no esta en el carro del usuario.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      tags: [cart]
      summary: Elimina un item del carro
      operationId: removeCartItem
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Carro actualizado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: El producto no estaba en el carro del usuario.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/ServerError"

  # ---------------------------------------------------------------- orders --
  /orders/quote:
    post:
      tags: [orders]
      summary: Calcula los totales del carro actual, sin confirmar la compra
      description: >
        Operacion idempotente, sin efectos secundarios (no toca stock, carro ni
        ordenes). Pensada para el paso "aplicar cupon" antes del checkout.
      operationId: quoteOrder
      security: [{ bearerAuth: [] }]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderQuoteRequest"
      responses:
        "200":
          description: Totales calculados.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderTotals"
        "400":
          description: >
            El carro esta vacio, o `couponCode` no existe, esta inactivo, expirado,
            o no alcanza el `minOrderAmount` del cupon.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
  /orders/checkout:
    post:
      tags: [orders]
      summary: Confirma la compra del carro actual
      description: >
        Revalida el stock (puede haber cambiado desde que se agrego al carro),
        descuenta stock, crea la orden (snapshot de items y precios, estado
        `pagado`) y vacia el carro — todo en una sola unidad de trabajo
        (Unit of Work): si algo falla, no se persiste ningun cambio parcial.
      operationId: checkout
      security: [{ bearerAuth: [] }]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderQuoteRequest"
      responses:
        "201":
          description: Orden creada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: El carro esta vacio, o el cupon enviado no es valido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Un producto del carro ya no existe en el catalogo.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: >
            Stock insuficiente para uno de los productos del carro al momento de
            pagar (condicion de carrera clasica entre agregar al carro y pagar).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/ServerError"
  /orders:
    get:
      tags: [orders]
      summary: Lista las ordenes del usuario autenticado
      operationId: listMyOrders
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Ordenes del usuario (vacia si no tiene ninguna).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
  /orders/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
        example: 1
    get:
      tags: [orders]
      summary: Obtiene una orden por id
      description: >
        **Nota de contrato:** la implementacion actual no filtra por
        propietario — cualquier usuario autenticado que conozca el `id` puede
        leer la orden. Documentado para que se evalue/pruebe explicitamente
        (candidato a control de autorizacion adicional).
      operationId: getOrderById
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Orden encontrada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

  # --------------------------------------------------------------- reviews --
...
  /admin/orders:
    get:
      tags: [admin-orders]
      summary: Lista todas las ordenes de todos los usuarios (admin)
      operationId: adminListOrders
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Todas las ordenes del sistema.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/ServerError"
  /admin/orders/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
        example: 1
    get:
      tags: [admin-orders]
      summary: Obtiene cualquier orden por id (admin)
      operationId: adminGetOrderById
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Orden encontrada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT devuelto por `POST /auth/register` o `POST /auth/login`. Payload:
        `{ id, email, role }`. Expira en 2 horas.
  responses:
    Unauthorized:
      description: Sin token (`Token no proporcionado`) o token invalido/expirado (`Token invalido o expirado`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Autenticado, pero sin rol admin (`Requiere rol admin`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: El recurso solicitado no existe.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ServerError:
      description: Error interno no controlado.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
      example: { error: "Mensaje descriptivo del error" }
...
    Installments:
      type: object
      properties:
        count: { type: integer, example: 12 }
        interestFree: { type: boolean, example: true }
      required: [count, interestFree]
    Product:
      type: object
      properties:
        id: { type: integer, example: 1 }
        sku: { type: string, example: SKU-001 }
        name: { type: string, example: Zapatillas Urban Runner }
        description: { type: string }
        price: { type: number, example: 45990 }
        listPrice: { type: number, example: 57490 }
        stock: { type: integer, example: 12 }
        category: { type: string, example: Calzado }
        brand: { type: string, example: UrbanStep }
        isNew: { type: boolean }
        installments:
          $ref: "#/components/schemas/Installments"
        images:
          type: array
          items: { type: string }
          description: Data URIs (SVG) de 3 variantes de color.
        videoUrl: { type: string, nullable: true }
        categoryPath:
          type: array
          items: { type: string }
          example: ["Inicio", "Calzado"]
        warrantyMonths: { type: integer, example: 12 }
        returnDays: { type: integer, example: 30 }
        imageUrl: { type: string, description: Data URI (SVG) de portada. }
      required:
        - id
        - sku
        - name
        - description
        - price
        - stock
        - category
        - imageUrl
...
    CartItem:
      type: object
      properties:
        productId: { type: integer }
        quantity: { type: integer }
        product:
          oneOf:
            - $ref: "#/components/schemas/Product"
            - type: "null"
        lineTotal: { type: number }
      required: [productId, quantity, product, lineTotal]
    CartResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
      required: [items]
    AddCartItemRequest:
      type: object
      properties:
        productId: { type: integer }
        quantity: { type: integer, minimum: 1 }
      required: [productId, quantity]
    SetCartItemQuantityRequest:
      type: object
      properties:
        quantity: { type: integer, minimum: 1 }
      required: [quantity]
    OrderQuoteRequest:
      type: object
      properties:
        couponCode:
          type: string
          nullable: true
          example: VERANO20
    OrderTotals:
      type: object
      properties:
        subtotal: { type: number, example: 45990 }
        discount: { type: number, example: 9198 }
        couponApplied:
          type: string
          nullable: true
          example: VERANO20
        tax: { type: number, description: IVA 19% sobre (subtotal - discount)., example: 6991 }
        shipping: { type: number, description: "0 si (subtotal - discount) >= 30000, si no 3000.", example: 0 }
        total: { type: number, example: 43783 }
      required: [subtotal, discount, couponApplied, tax, shipping, total]
    OrderItem:
      type: object
      description: Snapshot del producto al momento de la compra (no referencia al catalogo vivo).
      properties:
        productId: { type: integer }
        name: { type: string }
        unitPrice: { type: number }
        quantity: { type: integer }
      required: [productId, name, unitPrice, quantity]
    Order:
      type: object
      properties:
        id: { type: integer }
        userId: { type: integer }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal: { type: number }
        discount: { type: number }
        couponApplied:
          type: string
          nullable: true
        tax: { type: number }
        shipping: { type: number }
        total: { type: number }
        status:
          type: string
          description: 'Valor inicial siempre "pagado"; actualizable via `PUT /admin/orders/{id}/status`.'
          example: pagado
        createdAt: { type: string, format: date-time }
      required:
        - id
        - userId
        - items
        - subtotal
        - discount
        - couponApplied
        - tax
        - shipping
        - total
        - status
        - createdAt
...